refactor: ChessGame을 컨트롤러 레벨에 두고 뷰를 넘겨받아 스스로 출력하도록 수정
