+++ src/main/java/controller/ChessController.java
@@ -1,10 +1,8 @@
 package controller;
 
-import controller.command.Command;
-import domain.game.ChessGame;
-import view.InputView;
-import view.OutputView;
-import view.command.CommandType;
+import controller.command.*;
+import view.*;
+import view.command.*;
 
 public class ChessController {
     private final InputView inputView = new InputView();
@@ -22,9 +20,8 @@ public class ChessController {
         try {
             CommandType commandType = inputView.inputCommand();
             Command command = Command.from(commandType);
-            command.execute(chessGame);
+            command.execute(outputView, chessGame);
 
-            outputView.printChessBoard(chessGame.getChessBoard());
         } catch (final Exception exception) {
             OutputView.printErrorMessage(exception.getMessage());
             chessGame.end();
