+++ src/main/java/controller/ChessGame.java
@@ -1,7 +1,9 @@
-package domain.game;
+package controller;
 
 import controller.constants.*;
+import domain.game.*;
 import domain.position.*;
+import view.*;
 
 public class ChessGame {
     private final ChessBoard chessBoard;
@@ -12,15 +14,16 @@ public class ChessGame {
         gameState = GameState.NOT_STARTED;
     }
 
-    public void start() {
+    public void start(final OutputView outputView) {
         gameState = GameState.RUNNING;
+        outputView.printChessBoard(chessBoard);
     }
 
     public void end() {
         gameState = GameState.STOPPED;
     }
 
-    public void move(final Position source, final Position target) {
+    public void move(final OutputView outputView, final Position source, final Position target) {
         if (gameState.isNotStarted()) {
             throw new IllegalStateException("[ERROR] 게임이 시작되지 않았으므로 이동할 수 없습니다.");
         }
@@ -28,13 +31,10 @@ public class ChessGame {
             throw new IllegalStateException("[ERROR] 게임이 종료된 상태이므로 이동할 수 없습니다.");
         }
         this.gameState = chessBoard.move(source, target);
+        outputView.printChessBoard(chessBoard);
     }
 
     public boolean isRunning() {
         return gameState.isNotStarted() || gameState.isRunning();
-    }
-
-    public ChessBoard getChessBoard() {
-        return chessBoard;
     }
 }
